requirements.txt
cryptography>=41.0.0
psycopg2-binary==2.9.11
py-clob-client>=0.1.8
python-dotenv==1.2.1
requests>=2.31.0
thefuzz==0.22.1
